+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added Logger#handleExceptions to catch uncaught exceptions
 - added stack traces when an exception is logged
 - changed datetime to strftime, adds `%L` to output milliseconds
+- fixed Console handler from using accepting format options
 
 ## v0.2.0 - 2013-10-04